Add docker container for sampler.

+++ src/reservoir_sampler/sampler_manager.py
@@ -27,7 +27,8 @@ class SamplerManager(object):
         self.socket.bind(sampler_manager_address)
 
     def kill_sampler(self):
-        os.kill(self.sampler.pid, signal.SIGTERM)
+        if self.sampler is not None:
+            os.kill(self.sampler.pid, signal.SIGTERM)
 
     def handle_message(self, msg):
         print('Received request: %s' % msg)
